consolidate all the providers/decorators
