+++ src/SidebarSwitch.tsx
@@ -1,8 +1,9 @@
-import React from "react";
 import { SwitchProps } from "@/RoutingSwitch";
 import { Route, Switch } from "react-router-dom";
+import { useIsAuthenticated } from "@/providers/Profile";
 
-const SidebarSwitch: React.FC<SwitchProps> = ({ routes, authenticated }) => {
+const SidebarSwitch = ({ routes }: SwitchProps) => {
+    const authenticated = useIsAuthenticated();
     // Don't move the conditional inside render up to here (or a .filter)! It's
     // important the same <Route> are rendered here as for the main content, so
     // sidebars use the same scoping. Otherwise, you get weird 'exact' effects.
